Move probe function to util Update comment
